requirements.txt
Django==4.2.7
django-postgresql==0.0.3
djangorestframework==3.14.0
gunicorn==21.2.0
mysql-connector-python==8.2.0
psycopg2==2.9.9
dj_database_url==2.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.6.0